Lab 3

+++ INSSSIOK - Имплементација на системи со слободен и отворен код/Lab3/delete_product.php
@@ -1,11 +1,15 @@
 <?php
 
-include "database3.php";
+include 'db_connection.php';
+
 $db = getDatabaseConnection();
 
-$stmt = $db->prepare('DELETE FROM products WHERE id = :id');
-$stmt->bindValue(':id', $_GET['id']);
-$stmt->execute();
-
-header('Location: Lab3.php');
+if ($_SERVER['REQUEST_METHOD'] === 'POST' && isset($_POST['id'])) {
+    $stmt = $db->prepare("DELETE FROM products WHERE id = :id");
+    $id = $_POST['id'];
+    $stmt->bindValue(':id', $id, PDO::PARAM_INT);
+    $stmt->execute();
+    header('Location: index.php');
+}
 exit();
+?>
